Support datetime.date
Although the comment in nu-protocol/src/syntax_shape says DateTime can
be a date with no time, that doesn't appear to be so.

"Plugin failed to encode: input is not enough for unique date and time"

Add a midnight with an unrealistic timezone offset that we can
recognize and decode back to a datetime.date.